Rename flattening-ocds to flatten-tool
The Python module name must not have a dash in it, so I've used
flattentool.
