--- flatten-ocds
@@ -1,3 +1,0 @@
-#!/usr/bin/env python
-import flattening_ocds.cli
-flattening_ocds.cli.main()
