+++ flatten-tool
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+import flattentool.cli
+flattentool.cli.main()
