+++ flattentool/input.py
