+++ flattentool/tests/fixtures/release-schema.json
