+++ flattentool/tests/fixtures/tenders_releases_2_releases.json
