+++ flattentool/tests/fixtures/tenders_releases_base.json
