+++ setup.py
@@ -7,12 +7,12 @@ if sys.version < '3':
     install_requires.append('unicodecsv')
 
 setup(
-    name='flattening_ocds',
+    name='flattentool',
     version='0.0.0',
     author='Ben Webb',
     author_email='bjwebb67@googlemail.com',
-    packages=['flattening_ocds'],
-    scripts=['flatten-ocds'],
+    packages=['flattentool'],
+    scripts=['flatten-tool'],
     url='https://github.com/open-contracting/flattening-ocds',
     license='MIT',
     description='Tools for generating CSV and other flat versions of the structured data',
